backend/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
openai>=1.37.0
python-dotenv>=1.0.1
tldextract>=5.1.2
requests>=2.32.3

